models/speech_to_text/requirements.txt
ffmpeg-python==0.2.0
whisper.ai==1.0.0.1